playground/PG2/pyproject.toml
[project]
name = "pg2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "inngest>=0.5.8",
    "llama-index-core>=0.14.3",
    "llama-index-readers-file>=0.5.4",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
